+++ lib/xcodeproj/project.rb
@@ -115,13 +115,38 @@ module Xcodeproj
     #
     attr_reader :root_object
 
+    # A fast way to see if two {Project} instances refer to the same projects on
+    # disk. Use this over {#eql?} when you do not need to compare the full data.
+    #
+    # This shallow comparison was chosen as the (common) `==` implementation,
+    # because it was too easy to introduce changes into the Xcodeproj code-base
+    # that were slower than O(1).
+    #
+    # @return [Boolean] whether or not the two `Project` instances refer to the
+    #         same projects on disk, determined solely by {#path} and
+    #         `root_object.uuid` equality.
+    #
+    # @todo If ever needed, we could also compare `uuids.sort` instead.
+    #
+    def ==(other)
+      other && path == other.path && root_object.uuid == other.root_object.uuid
+    end
+
     # Compares the project to another one, or to a plist representation.
+    #
+    # @note This operation can be extremely expensive, because it converts a
+    #       `Project` instance to a hash, and should _only_ ever be used to
+    #       determine wether or not the data contents of two `Project` instances
+    #       are completely equal.
+    #
+    #       To simply determine wether or not two {Project} instances refer to
+    #       the same projects on disk, use the {#==} method instead.
     #
     # @param  [#to_hash] other the object to compare.
     #
     # @return [Boolean] whether the project is equivalent to the given object.
     #
-    def ==(other)
+    def eql?(other)
       other.respond_to?(:to_hash) && to_hash == other.to_hash
     end
 
@@ -131,9 +156,9 @@ module Xcodeproj
 
     alias_method :inspect, :to_s
 
-    # @return [Bool] Whether the xcproj conversion should be disabled. The
-    #         conversion can be disable also via the
-    #         `XCODEPROJ_DISABLE_XCPROJ` environment variable.
+    # @return [Boolean] Whether the `xcproj` conversion should be disabled. The
+    #         conversion can also be disabled via the `XCODEPROJ_DISABLE_XCPROJ`
+    #         environment variable.
     #
     attr_accessor :disable_xcproj
     def disable_xcproj?
@@ -174,9 +199,8 @@ module Xcodeproj
     def initialize_from_file
       pbxproj_path = path + 'project.pbxproj'
       plist = Xcodeproj.read_plist(pbxproj_path.to_s)
-      root_object_uuid = plist['rootObject']
       root_object.remove_referrer(self) if root_object
-      @root_object = new_from_plist(root_object_uuid, plist['objects'], self)
+      @root_object = new_from_plist(plist['rootObject'], plist['objects'], self)
       @archive_version =  plist['archiveVersion']
       @object_version  =  plist['objectVersion']
       @classes         =  plist['classes']