+++ lib/xcodeproj/project/object_attributes.rb
@@ -54,7 +54,7 @@ module Xcodeproj
         #   attribute.plist_name #=> projectRoot
         #
         def plist_name
-          CaseConverter.convert_to_plist(name, :lower)
+          @plist_name ||= CaseConverter.convert_to_plist(name, :lower)
         end
 
         # @return [Array<Class>] the list of the classes accepted by the